# src/main/java/it/unibo/sap/delivery/domain/deliveries/DeliveryStatus.java
package it.unibo.sap.delivery.domain.deliveries;

public enum DeliveryStatus {
    REQUESTED,
    VALIDATED,
    SCHEDULED,
    ASSIGNED,
    IN_PROGRESS,
    DELIVERED,
    REJECTED,
    CANCELLED,
    ABOLISHED
}